// functions/src/types/oauth/userInfo.d.ts
export interface OAuthUserInfo {
  sub: string;
  birthdate: string;
  family_name: string;
  gender: string;
  given_name: string;
  locale: string;
  middle_name: string;
  name: string;
  nickname: string;
  picture: string;
  preferred_username: string;
  profile: string;
  updated_at: number;
  website: string;
  zoneinfo: string;
  email: string;
  email_verified: boolean;
}